+++ extensions/elasticsearch/ActiveRecord.php
@@ -416,8 +416,9 @@ class ActiveRecord extends BaseActiveRecord
         $this->_version = $response['_version'];
         $this->_score = null;
 
+        $changedAttributes = array_fill_keys(array_keys($values), null);
         $this->setOldAttributes($values);
-        $this->afterSave(true, $values);
+        $this->afterSave(true, $changedAttributes);
 
         return true;
     }